[svn r2158] merged everything from my "merge" branch (BrianIsoGalaxies) into devel trunk.  Things that were added:
--"Galaxy formation" test problem, which allows one to simulate isolated
galaxies. Generously donated by Elizabeth Tasker.

--"Radiating shock" test problem, which is sedov+cooling/chemistry, and
also introduces an extended test problem generation formalism (allowing
test problems to have chemistry, metals, etc. etc.)

--"rotating cylinder" test problem, which simulates rotating, collapsing
cylinders of gas.  This is a good test of the isolated gravity test
parameters.

--re-added Gadget cooling.

--added metallicity-based refinement and the idea of a "must refine region", which
forces refinement to a specified level in a rectangular solid volume.
It is similar in concept to fixed nested grids, but doesn't require the creation
of actual grids.  Useful in varied circumstances.

--added extended point source gravity code in Grid:ComputeAccelerationFieldExternal.

--added parameter RootGridCourantSafetyNumber, which controls the timestep
taken on the root grid in order to avoid the unphysical ejection of stars from
subgrids (this is a silly bug).

--added the ability to set the particle-to-cell mass ratio by hand, which is
important when one is using parallel root grid IO and parallel particle IO
AND a particle-to-cell ratio that is not one.  I imagine nobody cares but me.