The start of a buffer handler for ActiveParticles.  This buffer handler will allocate internal buffers, and will be called by individual grids for IO and possibly communication, if it could work for this.
The idea would be in the IO to count the number of each type of active
particles in that grid.  Then allocate buffers for each type using the built-in
handler.  Then feed the hid_t of the group to each buffer handler.  Then,
de-allocate.  This may be feasible for communication as well, but that's not
quite in my wheelhouse.